requirements.txt
anyapi==1.1.3
argh==0.26.2
astroid==1.6.5
attrs==19.3.0
Babel==2.7.0
backcall==0.1.0
bcrypt==3.1.7
beautifulsoup4==4.6.0
bleach==3.1.0
blinker==1.4
certifi==2018.4.16
cffi==1.13.2
chardet==3.0.4
Click==7.0
colorama==0.3.9
cycler==0.10.0
decorator==4.4.1
defusedxml==0.6.0
docutils==0.14
entrypoints==0.3
Flask==1.1.1
Flask-BabelEx==0.9.3
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
idna==2.7
igql==1.1.0
importlib-metadata==0.23
ipykernel==5.1.3
ipython==7.9.0
ipython-genutils==0.2.0
isodate==0.6.0
isort==4.3.4
itsdangerous==1.1.0
jedi==0.15.1
Jinja2==2.10.3
joblib==0.14.0
jsonschema==3.1.1
jupyter-client==5.3.4
jupyter-core==4.6.1
jupyterthemes==0.20.0
lazy-object-proxy==1.3.1
lesscpy==0.13.0
lxml==4.4.2
MarkupSafe==1.1.1
matplotlib==3.1.1
mccabe==0.6.1
mistune==0.8.4
more-itertools==7.2.0
nbconvert==5.6.1
nbformat==4.4.0
notebook==6.0.1
numpy==1.17.3
pandas==0.25.3
pandocfilters==1.4.2
parso==0.5.1
passlib==1.7.1
pathtools==0.1.2
pickleshare==0.7.5
Pillow==6.1.0
pipenv==2018.11.26
pkgconfig==1.5.1
ply==3.11
prometheus-client==0.7.1
prompt-toolkit==2.0.10
py4j==0.10.8.1
pycparser==2.19
pycryptodome==3.6.6
Pygments==2.2.0
pylint==1.9.2
PyMySQL==0.9.3
pyparsing==2.4.2
pyrsistent==0.15.5
python-dateutil==2.8.0
pytz==2019.2
pywin32==225
pywinpty==0.5.5
PyYAML==5.1.2
pyzmq==18.1.0
requests==2.19.1
scikit-learn==0.21.3
scipy==1.3.2
Send2Trash==1.5.0
six==1.11.0
speaklater==1.3
SQLAlchemy==1.3.10
sqlparse==0.3.0
terminado==0.8.2
testpath==0.4.2
tornado==6.0.3
traitlets==4.3.3
urllib3==1.23
virtualenv==16.7.7
virtualenv-clone==0.5.3
watchdog==0.9.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
wrapt==1.10.11
WTForms==2.2.1
zipp==0.6.0
